+++ src/components/BackArrow/style.js
@@ -1,13 +1,14 @@
 import styled from 'styled-components';
+import { path } from 'ramda';
 
 export const ArrowDownIcon = styled.span`
   margin: auto;
   display: inline-block;
-  width: 10px;
-  height: 10px;
+  width: ${({ fontSize }) => (fontSize ? `calc(${fontSize} - (${fontSize}/2))` : '0.625rem')};
+  height: ${({ fontSize }) => (fontSize ? `calc(${fontSize} - (${fontSize}/2))` : '0.625rem')};
   background: transparent;
-  border-top: 2px solid var(--mainColor);
-  border-left: 2px solid var(--mainColor);
+  border-top: 2px solid ${({ color, theme }) => path(['color', color, 'mainColor'], theme)};
+  border-left: 2px solid ${({ color, theme }) => path(['color', color, 'mainColor'], theme)};
   text-decoration: none;
   color: transparent;
   transform: rotate(-45deg);
@@ -19,6 +20,6 @@ export const TextStyled = styled.p`
   font-style: normal;
   font-weight: 400;
   font-size: ${({ fontSize }) => fontSize || '1rem'};
-  line-height: 20px;
-  color: var(--mainColor);
+  line-height: ${({ fontSize }) => fontSize || '1rem'};;
+  color: ${({ color, theme }) => path(['color', color, 'mainColor'], theme)};
 `;